+++ packages/react/src/components/index.ts
@@ -1,4 +1,5 @@
 export * from './PostHogFeature'
+export * from './PostHogCaptureOnViewed'
 export {
     PostHogErrorBoundary,
     PostHogErrorBoundaryProps,