+++ packages/nextjs-config/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @posthog/nextjs-config
 
+## 1.3.10
+
+### Patch Changes
+
+- Updated dependencies [[`d8d98c9`](https://github.com/PostHog/posthog-js/commit/d8d98c95f24b612110dbf52d228c0c3bd248cd58)]:
+  - @posthog/core@1.5.1
+
+## 1.3.9
+
+### Patch Changes
+
+- Updated dependencies [[`068d55e`](https://github.com/PostHog/posthog-js/commit/068d55ed4193e82729cd34b42d9e433f85b6e606)]:
+  - @posthog/core@1.5.0
+
 ## 1.3.8
 
 ### Patch Changes